hyper-v framework: deploy 2nd VM to a different host
This adds a new option to deploy a VM on a secondary host. How this
works:
-- On a test configuration, if DeployOnSecondaryHyperVHost tag is
present and set to "yes", the VM will be deployed on the secondary host
specified in GlobalConfigurations.xml.
-- At runtime, the -TestLocation param has to be "host_1,host_2" for
this to be triggered.
Azure code not changed in this PR. Also, all the existing run methods on
Hyper-V are not affected