perf_fio: squash database results by qdepth
This patch changes how the fio performance results are uploaded
to the database, by squashing the results mode results by qdepth.

Previous behaviour was to insert a row in the database for each
mode (randread, randwrite, read, write) that have the same qdepth.

Current behaviour is to insert one row in the database for all
modes that have the same qdepth.